% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEFS_Paper_Functions.R
\name{obs.eff}
\alias{obs.eff}
\title{observer effects based on total abundance}
\usage{
obs.eff(
  data = NULL,
  ID = "ID",
  abundance = "abundance",
  year = "year",
  OE = 0.25,
  add.df = TRUE
)
}
\arguments{
\item{data}{dataframe including "abundance" per site "ID" and survey "year"
for all species (= raw data)}

\item{ID}{columname of site ID in \code{data}}

\item{abundance}{columname of abundance in \code{data}}

\item{year}{columname of year in \code{data}}

\item{OE}{threshold for classifying observer effects, should be a value >0
and <1.}

\item{add.df}{if TRUE, returns a dataframe with observer effect per site and
year}
}
\value{
\code{obs.eff()} returns a dataframe including the implemented raw data
(in "data") as well as the additionally column for observer effect ("OE_xx")
per site and year and the default column "abundance".

The function optionally returns the dataframe "df.oe" containing observer
effects per site and year if df.oe = TRUE.
This dataframe has the following columns:
\itemize{
\item ID:            site ID
\item year:          year of survey
\item Ab.total:      total abundance per site and year (summed up across all
species)
\item mean.Ab.total: mean total abundance (Ab.total) per site
\item Ab.prop:       proportional abundance per site and year
(= Ab.total / mean.Ab.total per site)
\item OE_xx:         observer effect levels "negative", "none", and "positive"
per site and year. \code{xx} is the used threshold in \%.
}
}
\description{
This function calculates observer effects based on the proportional total
abundance (= total abundance per site and year / mean total abundance per site).
For each site and year, observer effects are classified in three classes
using a defined threshold
\itemize{
\item if proportional abundance < (1-threshold), the effect is classified as
"negative".
\item if proportional abundance > (1+threshold), the effect is classified as
"positive".
\item if (1-threshold) <= proportional abundance <= (1+threshold), the effect is
classified as "none".
}
}
